All model view controller done

+++ admin/index.php
@@ -22,10 +22,6 @@
     $review = Review::all();
     $users = Users::all();
 
-    echo '<pre>'; 
-    var_dump($users);
-    echo '</pre>';
-
     $db = connectDB();
 
     $query = "SELECT * FROM reviews";
@@ -44,27 +40,36 @@
 
         $deleteID = filter_var($deleteID, FILTER_VALIDATE_INT);
 
-
-        echo '<pre>'; 
-        var_dump($_POST['image']);
-        echo '</pre>';
-
-        $upload_dir = '../uploads/';
-        
-        unlink($upload_dir . $_POST['image']);
-
         if ($deleteID) {
-            $review = Review::find($deleteID);
+        $type = $_POST['type'];
+       
+        if(valideContentType($type)){
 
+            if ($type === 'review') {
             
-            $review->delete();
-         
+                $upload_dir = '../uploads/';
+                
+                unlink($upload_dir . $_POST['image']);
+                $review = Review::find($deleteID);
+                $review->delete();
+            } else if ($type === 'user'){
+            
+                $users = Users::find($deleteID);
+                $users->delete();
+            }
+        }
         }
 
-        echo '<pre>'; 
-        var_dump($value2);
-        echo '</pre>';
+      
+        // if ($deleteID) {
+        //     $review = Review::find($deleteID);
 
+            
+        //     $review->delete();
+         
+        // }
+
+     
     }
 
     $resultQuery = mysqli_query($db, $query);
@@ -73,14 +78,14 @@
      includeTemplate("header", $header_text);
 
 
-  
-
 ?>
     
     <main class="container">
 
         <a href="/blog_sobreliteratura/admin/reviews/create.php" class="button-inline">Nueva reseña</a>
+        <a href="/blog_sobreliteratura/admin/users/create.php" class="button-inline">Nuevo usuario</a>
 
+        <h2>Reseñas</h2>
         <table class="reviews">
             <thead>
                 <tr>
@@ -114,11 +119,58 @@
                                 <input type="hidden" name="image" value="<?php echo s($value->image) ?>">
 
                                 <input type="hidden" name="id" value="<?php echo s($value->id) ?>">
+                                <input type="hidden" name="type" value="<?php echo 'review'; ?>">
 
                                 <input type="submit" class="button-block" value="Eliminar"> 
                             </form>
 
                             <a href="reviews/update.php?id=<?php echo $value->id ?>">
+                                <button class="button-block">Editar</button>
+                            </a>
+                        </td>
+                    </tr>
+                    <?php
+                }
+            
+        
+            ?>
+                </tbody>
+                
+        </table>
+        <h2 class="center">Escritores</h2>
+        <table class="reviews">
+            <thead>
+                <tr>
+                    <th>Id</th>
+                    <th>Nombre</th>
+                    <th>Apellido</th>
+                    <th>Acciones</th>
+                </tr>
+            </thead>
+            
+            <tbody>
+            <?php 
+            foreach ($users as $user) {
+                
+
+            // if ($result = $review->num_rows) {
+                // Loop through each row
+                    ?>
+                    <tr>
+                        <td><?php echo $user->id; ?></td>
+                        <td><?php echo $user->name; ?></td>
+                        <td><?php echo $user->lastname; ?></td>
+                        <td>
+                            <form method="POST">
+
+                                <input type="hidden" name="id" value="<?php echo s($user->id) ?>">
+
+                                <input type="hidden" name="type" value="<?php echo 'user'; ?>">
+
+                                <input type="submit" class="button-block" value="Eliminar"> 
+                            </form>
+
+                            <a href="users/update.php?id=<?php echo $user->id ?>">
                                 <button class="button-block">Editar</button>
                             </a>
                         </td>
